+++ packages/adapters/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tango-ts/adapters
 
+## 0.6.0
+
+### Patch Changes
+
+- @tango-ts/http@0.6.0
+
 ## 0.5.0
 
 ### Patch Changes